+++ app/Resources/views/base.html.twig
@@ -57,7 +57,7 @@
             {% if user is defined and user is not null %}
                 <li class="nav-item">
 
-                    <a class="nav-link" href="">Welcome {{ user.username }} LOGOUT </a>
+                    <a class="nav-link" href="{{ path('logout_Process') }}">Welcome {{ user.username }} LOGOUT </a>
                 </li>
             {% else %}
                 <li class="nav-item">
