--- src/InventoryBundle/Resources/views/Default/a1_order_history.html.twig
@@ -1,43 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block content %}
-    <div class="container">
-        <h1>A1 Order History</h1>
-        <hr/>
-        <table class="table">
-            <thead class="thead-inverse">
-            <tr>
-                <th>#</th>
-                <th>Date</th>
-                <th>Product Name</th>
-                <th>Quantity</th>
-                <th>Cost</th>
-            </tr>
-            </thead>
-            <tbody>
-            <tr>
-                <th scope="row">1</th>
-                <td></td>
-                <td></td>
-                <td></td>
-                <td></td>
-            </tr>
-            <tr>
-                <th scope="row">2</th>
-                <td></td>
-                <td></td>
-                <td></td>
-                <td></td>
-            </tr>
-            <tr>
-                <th scope="row">3</th>
-                <td></td>
-                <td></td>
-                <td></td>
-                <td></td>
-            </tr>
-            </tbody>
-        </table>
-    </div>
-
-{% endblock %}
