+++ src/InventoryBundle/Resources/views/Default/home.html.twig
@@ -3,5 +3,5 @@
 {% block content %}
 
     <div class="container">
-        <h1>Inventory Manager</h1>
+        <h1>Home</h1>
 {% endblock %}
