--- src/InventoryBundle/Resources/views/Default/show.html.twig
@@ -1,25 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block content %}
-
-    <div class="container">
-        <h1>Show Example</h1>
-        <table class="table">
-            <thead class="thead-inverse">
-            <tr>
-                <th>Product Name</th>
-                <th>Cost</th>
-            </tr>
-            </thead>
-            <tbody>
-            {% for item in inventory %}
-            <tr>
-                <td>{{ item.productname }}</td>
-                <td>{{ item.cost }}</td>
-            </tr>
-            {% endfor %}
-            </tbody>
-        </table>
-    </div>
-
-{% endblock %}
