+++ src/InventoryBundle/Resources/views/Inventory/edit.html.twig
