+++ src/InventoryBundle/Resources/views/Inventory/edit.inventory.form.html.twig
