+++ src/InventoryBundle/Resources/views/Inventory/index.html.twig
