+++ src/InventoryBundle/Resources/views/Store/stores.html.twig
